add 1 minute polling; move javascript to files and use a real minifier
Added support for a cheaper option, where the function is invoked
every minute but only polls each queue until the queue is empty.

Moved all the in-line javascript to dedicated files, and a build script
minifies them and injects them into the json template as it builds it
from the yaml.