Switch to cloudformation native YAML support
CloudFormation just added support for YAML, with some nice helper tags.
In order to keep parsing the YAML and using our !minify tag, we would have
had to implement all of the CF tags, which didn't sound fun, so now the
minimized code is injected directly into the YAML string in the build step.